Add script needed to publish epx-pop to AWS CloudArtifact
epx-pop is arguably legacy code at this point, and should be replaced
with functionality in epx-exec. However, some internal packages
(including COVID19-data which is required for autocal) were configured
to install epx-pop from JFrog previously. Publishing to CodeArtifact
allows us to continue using those packages in the same way as we
used to when installing on Cloud Runner.